Added support for OpenRC as an alternative to systemd

+++ pre-uninstall.sh
@@ -7,10 +7,11 @@ if [ "$EUID" -ne 0 ]
 fi
 
 HOME_DIR="$(eval echo "~$(logname)")"
+INIT_SYSTEM="systemd"
 
 # Argument parsing
 SHORT=h
-LONG=help
+LONG=help,openrc
 VALID_ARGS=$(getopt -a --options $SHORT --longoptions $LONG -- "$@")
 if [[ $? -ne 0 ]]; then
     exit 1;
@@ -19,6 +20,9 @@ fi
 eval set -- "$VALID_ARGS"
 while true; do
   case "$1" in
+    '--openrc')
+        INIT_SYSTEM="openrc"
+        ;;
     '--help' | '-h')
         echo "Usage: $0" 1>&2
         exit 0
@@ -31,11 +35,6 @@ while true; do
 done
 #
 
-SERVICES_DIR="./services"
-SERVICE_EXTENSION=".service"
-
-SERVICES="$(cd "$SERVICES_DIR" && find . -maxdepth 1 -maxdepth 1 -type f -name "*$SERVICE_EXTENSION" -exec basename {} "$SERVICE_EXTENSION" \;)"
-
 function sanitizePath() {
     local SANITIZED_PATH="$1"
     local SANITIZED_PATH=${SANITIZED_PATH//..\//}
@@ -44,12 +43,24 @@ function sanitizePath() {
     echo "$SANITIZED_PATH"
 }
 
-echo "disabling services"
-systemctl daemon-reload
-for SERVICE in $SERVICES ; do
-    SERVICE=$(sanitizePath "$SERVICE")
-    echo "stopping [$SERVICE]"
-    systemctl stop "$SERVICE" 2> "/dev/null" || true
-    echo "disabling [$SERVICE]"
-    systemctl disable "$SERVICE" 2> "/dev/null" || true
-done
+if [ "${INIT_SYSTEM}"  = "systemd" ]; then
+    SERVICES_DIR="./services"
+    SERVICE_EXTENSION=".service"
+
+    SERVICES="$(cd "$SERVICES_DIR" && find . -maxdepth 1 -maxdepth 1 -type f -name "*$SERVICE_EXTENSION" -exec basename {} "$SERVICE_EXTENSION" \;)"
+
+    echo "disabling services"
+    systemctl daemon-reload
+    for SERVICE in $SERVICES ; do
+        SERVICE=$(sanitizePath "$SERVICE")
+        echo "stopping [$SERVICE]"
+        systemctl stop "$SERVICE" 2> "/dev/null" || true
+        echo "disabling [$SERVICE]"
+        systemctl disable "$SERVICE" 2> "/dev/null" || true
+    done
+elif [ "${INIT_SYSTEM}" = "openrc" ]; then
+    echo "stopping fw-fanctrl"
+    rc-service fw-fanctrl stop
+    echo "diabling fw-fanctrl"
+    rc-update del fw-fanctrl default
+fi
